# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Atzicay API",
        "description": "API documentation for Atzicay",
        "contact": {
            "email": "atzicay@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "api/atzicay/v1/"
        }
    ],
    "paths": {
...
        "/game-sessions/{id}": {
            "get": {
                "tags": [
                    "GameSessions"
                ],
                "summary": "Get game session by ID",
                "description": "Retrieves a game session by its ID.",
                "operationId": "7c7298ee9d30c4e21804a577af29d2ff",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the game session",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Game session details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GameSession"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Game session not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "GameSessions"
                ],
                "summary": "Update an existing game session",
                "description": "Updates an existing game session.",
                "operationId": "9eed21341e1ffb84a15224b79f7cdd5f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the game session",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/GameSessionsDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Game session updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GameSession"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/game-sessions": {
            "post": {
                "tags": [
                    "GameSessions"
                ],
                "summary": "Create a new game session",
                "description": "Creates a new game session.",
                "operationId": "af78c7bca0807e413d1316d1f6b6f991",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/GameSessionsDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Game session created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GameSession"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/game-settings/{id}": {
            "get": {
                "tags": [
                    "GameSettings"
                ],
                "summary": "Get a game setting by ID",
                "description": "Retrieves a game setting by its ID.",
                "operationId": "8e6a7eef358d056bcc88d6667e5c97ba",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the game setting to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Game setting found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GameSettingDTO"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "GameSettings"
                ],
                "summary": "Update a game setting",
                "description": "Updates an existing game setting.",
                "operationId": "0dd728a9390ae2e59dda169f80c87bf0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the game setting to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/GameSettingDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Game setting updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GameSetting"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/game-settings": {
            "post": {
                "tags": [
                    "GameSettings"
                ],
                "summary": "Create a new game setting",
                "description": "Creates a new game setting.",
                "operationId": "7442942d40567006cc945d1abb1c29b3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/GameSettingDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Game setting created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GameSetting"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "GameSessionsDTO": {
                "title": "Game Sessions DTO",
                "description": "Data transfer object for game sessions",
                "properties": {
                    "ProgrammingGameId": {
                        "description": "ID of the programming game",
                        "type": "integer"
                    },
                    "StudentId": {
                        "description": "ID of the student",
                        "type": "integer"
                    },
                    "Duration": {
                        "description": "Duration of the game session in seconds",
                        "type": "integer"
                    },
                    "Won": {
                        "description": "Indicates if the game session was won",
                        "type": "boolean"
                    }
                },
                "type": "object"
            },
            "GameSettingDTO": {
                "title": "Game Setting DTO",
                "description": "Data transfer object for game settings",
                "properties": {
                    "GameInstanceId": {
                        "description": "The ID of the game instance",
                        "type": "integer"
                    },
                    "ConfigKey": {
                        "description": "The configuration key",
                        "type": "string"
                    },
                    "ConfigValue": {
                        "description": "The configuration value",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "GameSession": {
                "required": [
                    "id",
                    "programmingGameId",
                    "studentId",
                    "duration",
                    "won",
                    "dateGame"
                ],
                "properties": {
                    "Id": {
                        "type": "integer",
                        "example": 1
                    },
                    "ProgrammingGameId": {
                        "type": "integer",
                        "example": 123
                    },
                    "StudentId": {
                        "type": "integer",
                        "example": 456
                    },
                    "Duration": {
                        "type": "integer",
                        "example": 30
                    },
                    "Won": {
                        "type": "boolean",
                        "example": true
                    },
                    "DateGame": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-10-01T12:00:00Z"
                    }
                },
                "type": "object"
            },
            "GameSetting": {
                "required": [
                    "Id",
                    "GameInstanceId",
                    "ConfigKey",
                    "ConfigValue"
                ],
                "properties": {
                    "Id": {
                        "type": "integer",
                        "example": 1
                    },
                    "GameInstanceId": {
                        "type": "integer",
                        "example": 123
                    },
                    "ConfigKey": {
                        "type": "string",
                        "example": "background_color"
                    },
                    "ConfigValue": {
                        "type": "string",
                        "example": "#FFFFFF"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}